fix: change from post to get method and retrieve headers

+++ src/main/java/com/posmosalimos/geulgwi/api/tag/delete/controller/DelTagController.java
@@ -1,6 +1,9 @@
 package com.posmosalimos.geulgwi.api.tag.delete.controller;
 
 import com.posmosalimos.geulgwi.api.tag.delete.service.DelTagService;
+import com.posmosalimos.geulgwi.global.jwt.service.TokenManager;
+import com.posmosalimos.geulgwi.global.util.AuthorizationHeaderUtils;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
@@ -17,9 +20,19 @@ import org.springframework.web.bind.annotation.RestController;
 public class DelTagController {
 
     private final DelTagService delTagService;
+    private final TokenManager tokenManager;
 
     @DeleteMapping("/delete/{tagSeq}")
-    public ResponseEntity<Boolean> delete(@PathVariable("tagSeq") Long tagSeq) {
+    public ResponseEntity<Boolean> delete(@PathVariable("tagSeq") Long tagSeq,
+                                          HttpServletRequest httpServletRequest) {
+
+        String authorization = httpServletRequest.getHeader("Authorization");
+        AuthorizationHeaderUtils.validateAuthorization(authorization);
+
+        String accessToken = authorization.split(" ")[1];
+
+        tokenManager.validateToken(accessToken);
+
         delTagService.delete(tagSeq);
         log.info("tag - delete success");
         return ResponseEntity.ok(true);
